HSEARCH-3684 Move bootstrap mapping interfaces to the correct package
... with all the other bootstrap-only interfaces.
