+++ engine/src/main/java/org/hibernate/search/engine/common/impl/MappingBuildContextImpl.java
@@ -6,8 +6,8 @@
  */
 package org.hibernate.search.engine.common.impl;
 
-import org.hibernate.search.engine.mapper.mapping.spi.MappingBuildContext;
-import org.hibernate.search.engine.mapper.mapping.spi.MappingKey;
+import org.hibernate.search.engine.mapper.mapping.building.spi.MappingBuildContext;
+import org.hibernate.search.engine.mapper.mapping.building.spi.MappingKey;
 import org.hibernate.search.engine.reporting.spi.ContextualFailureCollector;
 
 class MappingBuildContextImpl extends DelegatingBuildContext implements MappingBuildContext {
