+++ engine/src/main/java/org/hibernate/search/engine/mapper/mapping/building/spi/MappingInitiator.java
@@ -6,9 +6,6 @@
  */
 package org.hibernate.search.engine.mapper.mapping.building.spi;
 
-import org.hibernate.search.engine.mapper.mapping.spi.MappingBuildContext;
-import org.hibernate.search.engine.mapper.mapping.spi.MappingPartialBuildState;
-
 /**
  * An object responsible for initiating a mapping by contributing its basic configuration (indexed types, type metadata),
  * then creating the mapper based on the configuration processed by the engine.
