+++ mapper/orm/src/main/java/org/hibernate/search/mapper/orm/mapping/impl/HibernateOrmMappingKey.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.orm.mapping.impl;
 
-import org.hibernate.search.engine.mapper.mapping.spi.MappingKey;
+import org.hibernate.search.engine.mapper.mapping.building.spi.MappingKey;
 import org.hibernate.search.mapper.orm.logging.impl.HibernateOrmEventContextMessages;
 
 import org.jboss.logging.Messages;
