+++ mapper/pojo/src/main/java/org/hibernate/search/mapper/pojo/mapping/spi/PojoMappingConfigurationContributor.java
@@ -6,7 +6,7 @@
  */
 package org.hibernate.search.mapper.pojo.mapping.spi;
 
-import org.hibernate.search.engine.mapper.mapping.spi.MappingBuildContext;
+import org.hibernate.search.engine.mapper.mapping.building.spi.MappingBuildContext;
 import org.hibernate.search.engine.mapper.mapping.building.spi.MappingConfigurationCollector;
 import org.hibernate.search.mapper.pojo.mapping.building.spi.PojoTypeMetadataContributor;
